some changes + make style + make quality
