+++ doctr/models/classification/__init__.py
@@ -4,4 +4,4 @@ from .vgg import *
 from .magc_resnet import *
 from .vit import *
 from .zoo import *
-from .textnetFast import *
+from .textnet_fast import *
