+++ doctr/models/modules/layers/tensorflow.py
@@ -1,17 +1,12 @@
-import math
-from typing import Any, Tuple
-
-import tensorflow as tf
 from tensorflow.keras import layers
 
 
 class RepConvLayer(layers.Layer):
-    def __init__(self, in_channels, out_channels, kernel_size, stride,  dilation=1, groups=1):
+    def __init__(self, in_channels, out_channels, kernel_size, stride, dilation=1, groups=1):
         super(RepConvLayer, self).__init__()
-        
-        padding = (int(((kernel_size[0] - 1) * dilation) / 2),
-                   int(((kernel_size[1] - 1) * dilation) / 2))
-        
+
+        padding = (int(((kernel_size[0] - 1) * dilation) / 2), int(((kernel_size[1] - 1) * dilation) / 2))
+
         self.activation = layers.ReLU()
         self.main_conv = layers.Conv2D(
             filters=out_channels,
@@ -21,49 +16,47 @@ class RepConvLayer(layers.Layer):
             dilation_rate=dilation,
             groups=groups,
             use_bias=False,
-            input_shape=(None, None, in_channels) 
+            input_shape=(None, None, in_channels),
         )
 
         self.main_bn = layers.BatchNormalization()
 
-        ver_pad = (int(((kernel_size[0] - 1) * dilation) / 2), 0)
-        hor_pad = (0, int(((kernel_size[1] - 1) * dilation) / 2))
-        
-        
+        (int(((kernel_size[0] - 1) * dilation) / 2), 0)
+        (0, int(((kernel_size[1] - 1) * dilation) / 2))
+
         if kernel_size[1] != 1:
             self.ver_conv = layers.Conv2D(
-            filters=out_channels,
-            kernel_size=(kernel_size[0], 1),
-            strides=(stride, 1),
-            padding='valid',  
-            dilation_rate=(dilation, 1),
-            groups=groups,
-            use_bias=False,
-            input_shape=(None, None, in_channels) 
+                filters=out_channels,
+                kernel_size=(kernel_size[0], 1),
+                strides=(stride, 1),
+                padding="valid",
+                dilation_rate=(dilation, 1),
+                groups=groups,
+                use_bias=False,
+                input_shape=(None, None, in_channels),
             )
             self.ver_bn = layers.BatchNormalization()
         else:
-             self.ver_conv, self.ver_bn = None, None
-             
+            self.ver_conv, self.ver_bn = None, None
+
         if kernel_size[0] != 1:
             self.hor_conv = layers.Conv2D(
                 filters=out_channels,
                 kernel_size=(1, kernel_size[1]),
                 strides=stride,
-                padding='valid',  # TensorFlow utilise 'valid' pour l'équivalent de 'same' de PyTorch
+                padding="valid",  # TensorFlow utilise 'valid' pour l'équivalent de 'same' de PyTorch
                 dilation_rate=dilation,
                 groups=groups,
                 use_bias=False,
-                input_shape=(None, None, in_channels)  # Spécifiez la forme de l'entrée ici
-                )
+                input_shape=(None, None, in_channels),  # Spécifiez la forme de l'entrée ici
+            )
             self.hor_bn = layers.BatchNormalization()
         else:
             self.hor_conv, self.hor_bn = None, None
-        
+
         self.rbr_identity = layers.BatchNormalization() if out_channels == in_channels and stride == 1 else None
 
     def forward(self, input):
-    
         main_outputs = self.main_bn(self.main_conv(input))
         vertical_outputs = self.ver_bn(self.ver_conv(input)) if self.ver_conv is not None else 0
         horizontal_outputs = self.hor_bn(self.hor_conv(input)) if self.hor_conv is not None else 0
